Tidy up code and refine the way shutdown / startup hooks are done with timeout for start up.
